optimize: make fake64 be ultra-bare-metal, hardcore the crap out of it
